Add --shell_wrapper_mode to simplify debugging of runpty
The WrapperMode is used for debugging the shell wrapper program. By
default it is set to `silent`, but it can also be set to `debug` or
`trace`. Debug will log wrapper internal events, such as signal
handling. Trace will also log the socket payload. When enabled in a
test case, a log file will be created per shell on the extra logs
directory.